+++ context_processors.py
@@ -2,12 +2,10 @@ from .models import Data
 
 
 def seo(request):
-	try:
-		seo_data = Data.objects.get(public=True, url=request.path_info, site=request.site)
-	except:
-		seo_data = False
-	return {
-		'seo': seo_data,
-		'site': request.site,
-		'url': request.url
-	}
+    try:
+        seo_data = Data.objects.get(public=True, url=request.path_info, site=request.site)
+    except:
+        seo_data = False
+    return {
+        "seo": seo_data,
+    }