+++ urls.py
@@ -1,7 +1,9 @@
-from django.conf.urls import url
+from django.urls import path
 
 from . import views
 
+
 urlpatterns = [
-	url(r'^robots\.txt$', views.robots, name='robots'),
+    path("robots.txt", views.robots_txt, name="robots_txt"),
+    path("sitemap.xml", views.sitemap_xml, name="sitemap_xml"),
 ]